Plank Shake Animation suggestion
Just a suggestion as there's really nothing bad about how the planks currently animate.
This is only applied to plank1, the center "Ack" plank.
This code uses the same math to shake the planks as tilting the barrels. At a fast speed, it pretty much looks the same as the way it's currently animated, but this way more things are tweakable.
How far from center is moves (distance)
How many times it wobbles back and forth before coming to a rest (stiffness)
How fast it moves (speed)
This would give you the option to make it shake it dynamically, such as only a little shake if the ball barely touches it, or a huge shake if the ball hits it a full speed. Also, you'd have the option do a slow motion thing, say for a special monkey mode where things just move slow?