update3
added kickers for voodoo and the bar so the code can be written. Did some grapics, added a shark and parrot. did the spinner graphics. 
added lane lights li006 till li013 
added the code of the beard and the code of the trees.

Issues: trees glitch and the beard goes to high trough the pirate