Let's get Voodoo started
Added a bunch of stuff to start getting some modes started.
Right now if you hit the three insult targets, it will light the voodoo shot.  You can then hit the voodoo shot to start that.
Effects and what-not are in, but there is space for it.  As always, please let me know of any concerns.