voodoo mode start changes
Shooting the voodoo lady when not in a mode, the ball will hold for 1 second to do whatever is you want to do.
Shooting the voodoo lady when she's ready will hold the ball while the wheel spins (right now it's set for the dial-a-pirate but we can change that).  When the wheel finishes spinning, the proper voodoo mode will start, the ball hold for another 2 seconds, then be released.
Shooting the voodoo lady while in a voodoo mode will hold the ball for 2 seconds to give the player something, then release.