Some code tweaks, fixing test balls, and we'll need to talk about the lower kickout
As is commented (and as always, feel free to uncomment before/after committing), you don't need two sub to pick the direction Le Chuck shoots out.  But, if you feel it makes more sense your way, just keep your code.
The main point of this was, now that you're destroying balls and recreating them elsewhere, generally speaking it's a good idea to keep ball IDs, if you use them.  I don't think you do, but I do for ball testing.  The way the ball testing works is, when the test kicker creates a ball, it gives an ID of 21.  When the ball drains, if the ball it destroys has an ID of 21, it immediately exits the drain sub.  So we needed to put the ball ID on the test balls, otherwise they weren't work.  It's not a perfect solution because right now, if you do something to make a ball kick out of kicker002 before the Voodoo lady, when that ball drains it will cause a regular end of ball.  So just make sure you push ' to shoot the voodoo lady before doing anything else, test wise.
Also though, the way kicker002 is currently implemented, things could get nasty if two balls come out of it too quickly.  I think using a real timer, instead of a virtual vpmtimer would fix it, but we could talk about that later.
Otherwise, awesome update.