Pirate Ship suggestion
This is just a suggestion.  It is kind of buggy at the moment, but I didn't want to waste time make it perfect if you didn't like the idea.
Just let me know if you want me to continue working on this or not.
Maybe we could incorporate this into a skill shot?
If you wanted, we could make the boat rock, like it's on the ocean, the whole time?